+++ customize_sale/models/__init__.py
@@ -1,5 +1,6 @@
 # See LICENSE file for full copyright and licensing details.
 
+from . import sale_promotion
 from . import sale_order
 from . import sale_order_line
 from . import stock_picking
