+++ customize_sale/models/sale_promotion.py
