--- sale_promotion/models/__init__.py
@@ -1,2 +1,0 @@
-from . import sale_promotion
-from . import sale_order
